+++ Final Assignment/Final Assignment_DOUBLE/A2Matrix_DOUBLE/equal_DOUBLE.c
@@ -2,7 +2,7 @@
 #include "A2Matrix_DOUBLE.h"
 int eql(double A, double B)
 {
-    int E=0.0000001; // E is the precision
+    double E=0.0000001; // E is the precision
 
     if(fabs(A-B)<=E) return 1;
     else return 0;
